Implement authorization using Policy

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Gate;
+
 
 class QuestionsController extends Controller
 {
@@ -13,6 +15,12 @@ class QuestionsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+    
     public function index()
     {
         $questions = Question::with('user')->latest()->paginate(5);
@@ -51,7 +59,8 @@ class QuestionsController extends Controller
      */
     public function show(Question $question)
     {
-      //
+        $question->increment('views');
+        return view('questions.show',compact('question'));
     }
 
     /**
@@ -61,7 +70,13 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
-    {
+    {   //Gate method for authorization,, where in index.blade file set @if(Auth::user()->can('update-question', $question))
+        /*if(Gate::denies('update-question',$question)){
+            abort(403,'Access denied');   
+        }
+        return view('questions.edit', compact('question'));*/
+
+        $this->authorize("update",$question);
         return view('questions.edit', compact('question'));
     }
 
@@ -73,7 +88,13 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(AskQuestionRequest $request, Question $question)
-    {
+    {   //Gate method for authorization, where in index.blade file set @if(Auth::user()->can('update-question', $question))
+        /*if(Gate::denies('update-question',$question)){
+            abort(403,'Access denied');   
+        }$question->update($request->only('title','body'));
+        return redirect('/questions')->with('success',"Your question has been updated.");*/
+
+        $this->authorize("update",$question);
         $question->update($request->only('title','body'));
         return redirect('/questions')->with('success',"Your question has been updated.");
     }
@@ -84,8 +105,15 @@ class QuestionsController extends Controller
      * @param  \App\Question  $question
      * @return \Illuminate\Http\Response
      */
+
     public function destroy(Question $question)
-    {
+    {   //Gate method for authorization, where in index.blade file set @if(Auth::user()->can('delete-question', $question))
+        /*if(Gate::denies('delete-question',$question)){
+            abort(403,'Access denied');   
+        }$question->delete();
+        return redirect('/questions')->with('success',"Your question has been deleted.");*/
+
+        $this->authorize("delete",$question);
         $question->delete();
         return redirect('/questions')->with('success',"Your question has been deleted.");
     }
